Fix 2 race issues on handover between IWLAN and WWAN
1. If handover needs to revert back immediately, this may happen during
the previous one is ongoing, so that it is ignored because the current
transport is the same as the final transport.

2. If the previous EVENT_SETUP_DATA_CONNECTION_DONE isn't yet processed,
TNF can't find the data connection of original transport as it is
activating, then setup call on new transport won't be originated, and
the orignal data connection will be released.

Change-Id: I5ca905d353bd3d5ecf7ca8d0824b3607dca32138
CRs-Fixed: 2768688